Fixed New Conversation Display
When clicking 'Message' from a Professional Page and being redirected to the Dashboard Messages page, the subsequently created or fetched conversation was being filtered out from the results as it contained no messages.

The conversation from the URL parameter is now passed to the backend function so that this conversation can be included in the final list of conversations for the user in the UI. This ensures that a user can immediately begin typing a new message.

(Note that this is only in the case where the conversation URL parameter is present. If not, the conversation between the two users will be filtered out as normal.)