Added Cookie Consent Popup
Created Cookie Consent Popup Component and added to the Root Template. The component contains cookie consent details, an Accept button and a Learn More button.

Added Cookie Consent column to user table via database migration, added an action to set the cookie consent, and included this field when fetching user information.

The functionality should check the database for prior consent if the user is logged in, as well as local storage for the current session if the user is logged out. In these cases the popup will not be shown, but will be shown if the user has not yet given consent.

Additionally, the cookie consent field in the database is updated accordingly if a logged-out user subsequently signs up or logs in.