Added Post Verification Redirect URL Passthrough
A redirectTo Prop has been added to the SignUpModal and propagated through the Sign Up and Email Verification Flow so that the user is redirected to the specified page after clicking the verification link in the email.

When clicking the Book Now button, the intended booking page URL is passed so that the user can continue booking seamlessly.

Also tidied up the Submit handler so that a redirect is only performed to the booking page if the user has successfully logged in.