Fixed Non-Subscribed Professional Services Booking
Added fetching of the professional's profile for the Profile Services Page to check if the professional is subscribed, and used this Boolean as part of the condition for showing the Book Now button.

Also added null return when fetching the service for a booking page when the professional is not subscribed, which should trigger the redirect to notFound.