Resolved userExistsByEmail Issues
Removed Public Key Fallback as this cannot be used with the Admin API. Updated error messages to be generic, with error details being logged to the server only.

Fixed user check to use the listUsers function, as querying the auth.users table is not possible and returns an error. The action now uses a loop to iterate through paginated results if necessary and make the email comparison using a code filter. This change has been made due to a limitation with the Supabase Admin API.