Dynamic Date Range Data Refetching & Graph Implementation
Updated the Date Range Picker implementation and refactored the Admin Dashboard so that it supports dynamic data re-fetching based on the selected dates. This included a wrapper and Context Provider, so that the date range can be used by all tabs/pages in the dashboard consistently.

Added rechart and implemented the Booking Activity Graph.

Adjusted the Overview page UI to be more aligned.

Added a new date formatting function in the formatDate util.

Renamed AdminDashboardPage & Template files to AdminOverviewPage & Template files to avoid confusion. The AdminDashboardPageLayout Remains as is.