Updated Closest Location Sorting
Updated the closest location sorting to use an actual distance calculation between the supplied location and the coordinates of the returned services.

Added default location which fetches the user's saved address from the database, based on their client or professional profile. For logged out users, or users with no address saved, the function defaults to using the user's current location via the Geolocation API.

This is passed to the sorting function to sort the results accurately by closest distance when Closest Location is selected.