Updated Admin Invite & Appointments Fetching
Updated the workflow for inviting a new admin to correctly set the user's role (DB migration adjusting the new user function), and send a password reset email for them to set their password.

Refactored the Appointments fetching to utilise nested select statements instead of multiple database calls.

Updated the Invite Admin modal to reset on close for ease of inviting multiple admins in a row.