Added Professional Profile Admin Modal + Max Services
Added a modal to display professional profile details, including services and appointments, accessible from the Admin Professionals tab by clicking on a Professional entry.

Added a Max Services column in the database for Professional Profiles, which can be set in the new modal (updates the database on input change).

Updated the services limit logic for the Professional Profile page so that it uses the new field, and defaults to the admin default if null, or 50 if both null.

The Max Services column should be automatically populated from the Admin Default when a new professional profile is created (new user sign up).