Refactored & Fixed Timeslot Selection & Highlighting For Bookings
Rewrote the timeslot and availability highlighting to be correct to the length of time of the service and not allow overlapping selections with existing appointments. The duration calculations now use a time-based approach instead of the previous timeslots-based approach.

Highlighting should now not highlight non-consecutive timeslots, so that new appointments up to existing appointments can be booked and this will be shown correctly. The end time calculation has also been resolved so that it shows the correct time in that scenario.

Front-end timeslot disabling has been removed as we are already performing the availability timeslot filtering on the backend for existing appointments.