Refactored Invite Admin & Admins Data Fetching
Removed the API endpoint for inviting an admin and refactored to use a server action directly instead.

Moved all sorting and filtering to be server-side in the server action.

Added path revalidation and end date adjustment to the admin invitation function to ensure new admins are displayed immediately after invitation.