Resolved Admin Support Request Message Fetching
Added function to fetch support request messages for an admin user viewing the support request details, and integrated this into the Support Request Detail Page and Chat Widget.

Updated the fetching of the message sender's avatar/initials to be correct for a viewing admin user.

Added a helper function for fetching user names to display in the People Involved section.

Removed 'Support Request' button from Booking Details page when an admin is viewing.

Changed the 'Back To Appointments' link to redirect to /admin/appointments when an admin user clicks it.